Authenticate the vowel skewer
Vowel Skewers
An authentic vowel skewer is a skewer with a delicious and juicy mix of consonants and vowels. However, the way they are made must be just right:

Skewers must begin and end with a consonant.

Skewers must alternate between consonants and vowels.

There must be an even spacing between each letter on the skewer, so that there is a consistent flavour throughout.

Create a function which returns whether a given vowel skewer is authentic.